#2 set ID as optional, id is autogenerated by database

+++ src/main/kotlin/pl/szymonleyk/betgame/register/account/AccountRepository.kt
@@ -4,4 +4,4 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface AccountRepository : ReactiveCrudRepository<Account, Long>
+interface AccountRepository : ReactiveCrudRepository<Account, Int>
