+++ src/main/kotlin/pl/szymonleyk/betgame/wallettransactions/WalletTransactionRepository.kt
@@ -4,4 +4,4 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface WalletTransactionRepository : ReactiveCrudRepository<WalletTransaction, Long>
+interface WalletTransactionRepository : ReactiveCrudRepository<WalletTransaction, Int>
